Début gestion des erreurs

+++ src/main/java/com/droovy/request/UserApiRequest.java
@@ -25,8 +25,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sun.jersey.api.client.ClientResponse.Status;
 
-import errors.ApplicationException;
-import errors.UserFaultException;
+import errors.UserApplicationError;
 
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 
@@ -41,8 +40,12 @@ public class UserApiRequest {
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/list")
-	public Response getFilesList(@Context UriInfo uriInfo,@QueryParam("path") String path,@QueryParam("idUser") String idUser,@QueryParam("idFolder") String idFolder,@QueryParam("getDropbox") int getDropbox,@QueryParam("getGoogleDrive") int getGoogledrive,@QueryParam("getOnedrive") int getOnedrive) throws JsonProcessingException, ApplicationException, UserFaultException {
+	public Response getFilesList(@Context UriInfo uriInfo,@QueryParam("path") String path,@QueryParam("idUser") String idUser,@QueryParam("idFolder") String idFolder,@QueryParam("getDropbox") int getDropbox,@QueryParam("getGoogleDrive") int getGoogledrive,@QueryParam("getOnedrive") int getOnedrive) throws JsonProcessingException {
 				
+		if(idUser == null || path==null || idFolder == null) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
 		List<File> listDropbox = new LinkedList<>(), listGoogleDrive = new LinkedList<>(),listOneDrive = new LinkedList<>();
 		
 		if(getDropbox==1) {			
@@ -81,11 +84,16 @@ public class UserApiRequest {
 
 	
 	@POST
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Consumes(MediaType.MULTIPART_FORM_DATA)
 	@Path("/upload")
-	public String uploadFile(@FormDataParam("file") InputStream uploadInputStream, @FormDataParam("file") FormDataContentDisposition fileDetail, @FormDataParam("idUser") String idUser, @FormDataParam("drive") String drive,@FormDataParam("pathInDrive") String pathInDrive) throws IOException {
+	public Response uploadFile(@FormDataParam("file") InputStream uploadInputStream, @FormDataParam("file") FormDataContentDisposition fileDetail, @FormDataParam("idUser") String idUser, @FormDataParam("drive") String drive,@FormDataParam("pathInDrive") String pathInDrive,@FormDataParam("parentId") String parentId) throws IOException {
 	
+		if(idUser == null || uploadInputStream==null || fileDetail == null || drive == null || parentId == null || pathInDrive == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		
 		OutputStream outputStream = new FileOutputStream(new java.io.File(fileDetail.getFileName()));
 	
 		
@@ -100,25 +108,37 @@ public class UserApiRequest {
 		outputStream.close();
 		uploadInputStream.close();		
 		
+		File uploadedFile;
+		
 		if(drive.equals("dropbox")) {
-			request_dropbox.uploadFile(fileDetail.getFileName(),pathInDrive, idUser);
+			uploadedFile=request_dropbox.uploadFile(fileDetail.getFileName(),pathInDrive, idUser,parentId);
 		}
 		else if(drive.equals("onedrive")) {
-			request_onedrive.uploadFile(fileDetail.getFileName(), pathInDrive, idUser);
+			uploadedFile=request_onedrive.uploadFile(fileDetail.getFileName(), pathInDrive, idUser,parentId);
 		}
 		else if(drive.equals("googledrive")) {
-			request_googledrive.uploadFile(fileDetail.getFileName(), pathInDrive, idUser);
+			uploadedFile=request_googledrive.uploadFile(fileDetail.getFileName(), pathInDrive, idUser,parentId);
+		}
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
 		}
 		
 		
+		ObjectMapper mapper = new ObjectMapper();
+		String output = "{" + mapper.writeValueAsString(uploadedFile)+"}";
 		
-		return fileDetail.getFileName()+" "+fileDetail.getSize()+" "+fileDetail.getType()+" "+fileDetail;
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
 	}
 
 	@GET
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/delete")
-	public String deleteFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive) throws IOException {
+	public Response deleteFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive) throws IOException {
+		
+		if(idUser == null || path==null || idFile == null || drive == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
 		
 		if(drive.equals("dropbox")) {
 			request_dropbox.removeFile(idFile, path, idUser);
@@ -129,61 +149,101 @@ public class UserApiRequest {
 		else if(drive.equals("googledrive")) {
 			request_googledrive.removeFile(idFile, path, idUser);
 		}
-		return "";
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity("{}").build();
 	}
 	
 	@GET
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/rename")
-	public String renameFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("name") String name) throws IOException {
+	public Response renameFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("name") String name) throws IOException {
 
-		if(drive.equals("dropbox")) {
-			request_dropbox.renameFile(idFile, path, name, idUser);
+		File renamedFile;
+		
+		if(idUser == null || path==null || idFile == null || drive == null || name == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
 		}
-		else if(drive.equals("onedrive")) {
-			request_onedrive.renameFile(idFile, path, name, idUser);
-		}
-		else if(drive.equals("googledrive")) {
-			request_googledrive.renameFile(idFile, path, name, idUser);
-		}
-		return "";
-	}
-	
-	@GET
-	@Produces("text/plain")
-	@Path("/move")
-	public String moveFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("idParent") String idParent, @QueryParam("pathParent") String pathParent, @QueryParam("name") String name) throws IOException {
-
-		if(drive.equals("dropbox")) {
-			request_dropbox.moveFile(idFile, path, idParent, pathParent, idUser,name);
-		}
-		else if(drive.equals("onedrive")) {
-			request_onedrive.moveFile(idFile, path, idParent, pathParent, idUser,name);
-		}
-		else if(drive.equals("googledrive")) {
-			request_googledrive.moveFile(idFile, path, idParent, pathParent, idUser,name);
-		}
-		return "";
-	}
-	@GET
-	@Produces("text/plain")
-	@Path("/freespace")
-	public String freeSpace( @QueryParam("idUser") String idUser, @QueryParam("drive") String drive) throws IOException {
+		
 		
 		if(drive.equals("dropbox")) {
-			return request_dropbox.freeSpaceRemaining(idUser);
+			renamedFile=request_dropbox.renameFile(idFile, path, name, idUser);
 		}
 		else if(drive.equals("onedrive")) {
-			return request_onedrive.freeSpaceRemaining(idUser);
+			renamedFile=request_onedrive.renameFile(idFile, path, name, idUser);
 		}
 		else if(drive.equals("googledrive")) {
-			return request_googledrive.freeSpaceRemaining(idUser);
+			renamedFile=request_googledrive.renameFile(idFile, path, name, idUser);
 		}
-		return "{}";
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		ObjectMapper mapper = new ObjectMapper();
+		String output = "{" + mapper.writeValueAsString(renamedFile)+"}";
+			
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
+	}
+	
+	@GET
+	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/move")
+	public Response moveFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("idParent") String idParent, @QueryParam("pathParent") String pathParent, @QueryParam("name") String name) throws IOException {
+
+		File movedFile;
+
+		if(idUser == null || path==null || idFile == null || drive == null || name == null || pathParent == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		
+		if(drive.equals("dropbox")) {
+			movedFile = request_dropbox.moveFile(idFile, path, idParent, pathParent, idUser,name);
+		}
+		else if(drive.equals("onedrive")) {
+			movedFile = request_onedrive.moveFile(idFile, path, idParent, pathParent, idUser,name);
+		}
+		else if(drive.equals("googledrive")) {
+			movedFile = request_googledrive.moveFile(idFile, path, idParent, pathParent, idUser,name);
+		}
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		ObjectMapper mapper = new ObjectMapper();
+		String output = "{" + mapper.writeValueAsString(movedFile)+"}";
+			
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
+	}
+	@GET
+	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/freespace")
+	public Response freeSpace( @QueryParam("idUser") String idUser, @QueryParam("drive") String drive) throws IOException {
+		
+		if(idUser == null || drive == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		String output;
+		
+		if(drive.equals("dropbox")) {
+			output=request_dropbox.freeSpaceRemaining(idUser);
+		}
+		else if(drive.equals("onedrive")) {
+			output=request_onedrive.freeSpaceRemaining(idUser);
+		}
+		else if(drive.equals("googledrive")) {
+			output=request_googledrive.freeSpaceRemaining(idUser);
+		}
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
 			
 	}
 	
-	@GET
+	/*@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/share")
 	public Response shareFile(@QueryParam("folder") String folder,@QueryParam("permission") String permission, @QueryParam("message") String message,@QueryParam("idUser") String idUser,@QueryParam("idFile") String idFile,@QueryParam("mail") String mail){
@@ -198,16 +258,21 @@ public class UserApiRequest {
 		
 		return Response.status(Status.OK).entity("").build();
 		
-	}
+	}*/
+	
+	
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/search")
 	public Response searchFile(@QueryParam("idUser") String idUser,@QueryParam("query") String query,@QueryParam("getDropbox") int getDropbox,@QueryParam("getGoogleDrive") int getGoogledrive,@QueryParam("getOnedrive") int getOnedrive) throws JsonProcessingException{
 	
-
+		if(idUser == null || query == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
 		List<File> listDropbox = new LinkedList<>(), listGoogleDrive = new LinkedList<>(),listOneDrive = new LinkedList<>();
 		
-		if(getDropbox==1) {			//Transformer en fonction ou hasmpa qui regarde si token null
+		if(getDropbox==1) {			//Transformer en fonction ou hashmap qui regarde si token null
 			listDropbox = request_dropbox.searchFile(idUser,query);
 		}
 		if(getGoogledrive==1) {
